+++ clang/include/clang/Lex/Lexer.h
@@ -185,12 +185,12 @@ public:
   Lexer(const Lexer &) = delete;
   Lexer &operator=(const Lexer &) = delete;
 
-  /// Create_PragmaLexer: Lexer constructor - Create a new lexer object for
-  /// _Pragma expansion.  This has a variety of magic semantics that this method
-  /// sets up.
-  static std::unique_ptr<Lexer> Create_PragmaLexer(
-      SourceLocation SpellingLoc, SourceLocation ExpansionLocStart,
-      SourceLocation ExpansionLocEnd, unsigned TokLen, Preprocessor &PP);
+  /// CreateScratchLexer: Lexer constructor - Create a new lexer object that
+  /// places 'Code' in the scratch buffer and lexes from it. This is used to
+  /// support '_Pragma' and string injection.
+  static std::unique_ptr<Lexer>
+  CreateScratchLexer(StringRef Code, SourceLocation ExpansionLocStart,
+                     SourceLocation ExpansionLocEnd, Preprocessor &PP);
 
   /// getFileLoc - Return the File Location for the file we are lexing out of.
   /// The physical location encodes the location where the characters come from,