+++ clang/lib/Sema/SemaExpand.cpp
@@ -76,6 +76,26 @@ static auto InitListContainsPack(const InitListExpr *ILE) {
                       [](const Expr *E) { return isa<PackExpansionExpr>(E); });
 }
 
+static Token CreateValueDeclAnnotToken(ValueDecl *D, SourceLocation Loc) {
+  Token Tok;
+  Tok.startToken();
+  Tok.setKind(tok::annot_value_decl);
+  Tok.setAnnotationValue(D);
+  Tok.setAnnotationEndLoc(Loc);
+  Tok.setLocation(Loc);
+  return Tok;
+}
+
+static Token CreateExprAnnotToken(Expr *E, SourceLocation Loc) {
+  Token Tok;
+  Tok.startToken();
+  Tok.setKind(tok::annot_expr);
+  Tok.setAnnotationValue(ExprResult(E).getAsOpaquePointer());
+  Tok.setAnnotationEndLoc(Loc);
+  Tok.setLocation(Loc);
+  return Tok;
+}
+
 static bool HasDependentSize(const DeclContext *CurContext,
                              const CXXExpansionStmtPattern *Pattern) {
   switch (Pattern->getKind()) {
@@ -203,11 +223,17 @@ static IterableExpansionStmtData TryBuildIterableExpansionStmtInitializer(
   if (BeginStmt.isInvalid())
     return Data;
 
-  // TODO: Build 'constexpr auto iter = begin + decltype(begin - begin){i};'.
-  S.Diag(ColonLoc, diag::err_iterating_expansion_stmt_unsupported);
-  return Data;
+  // Build 'begin + decltype(begin - begin){i}'.
+  ExprResult BeginPlusI = S.TokenInjectionHandler->ParseAsExpression(
+      "__begin + decltype(__begin - __begin){__i}",
+      {
+          {"__begin", CreateValueDeclAnnotToken(Info.BeginVar, ColonLoc)},
+          {"__i", CreateExprAnnotToken(Index, ColonLoc)},
+      },
+      ColonLoc);
+  if (BeginPlusI.isInvalid())
+    return Data;
 
-#if 0 // This will be used once we support iterating expansion statements.
   // Store it in a variable.
   // See also Sema::BuildCXXForRangeBeginEndVars().
   const auto DepthStr = std::to_string(Scope->getDepth() / 2);
@@ -233,7 +259,6 @@ static IterableExpansionStmtData TryBuildIterableExpansionStmtInitializer(
   Data.IterDecl = IterVarStmt.getAs<DeclStmt>();
   Data.TheState = IterableExpansionStmtData::IsIterableResult::Iterable;
   return Data;
-#endif
 }
 
 static StmtResult BuildDestructuringDecompositionDecl(
@@ -622,11 +647,35 @@ Sema::ComputeExpansionSize(CXXExpansionStmtPattern *Expansion) {
     EnterExpressionEvaluationContext ExprEvalCtx(
         *this, ExpressionEvaluationContext::ConstantEvaluated);
 
-    // TODO: Build the lambda and evaluate it.
-    Diag(Loc, diag::err_iterating_expansion_stmt_unsupported);
-    return std::nullopt;
+    // Annotation token that instructs the parser to declare begin/end, i.e.
+    //
+    //   auto __begin = begin-expr;
+    //   auto __end = end-expr;
+    //
+    Token DeclareBeginEnd;
+    DeclareBeginEnd.startToken();
+    DeclareBeginEnd.setKind(tok::annot_expansion_stmt_declare_begin_end);
+    DeclareBeginEnd.setAnnotationValue(Expansion->getRangeVar());
+    DeclareBeginEnd.setLocation(Loc);
+    DeclareBeginEnd.setAnnotationEndLoc(Loc);
 
-#if 0 // This will be used once we support iterating expansion statements.
+    // Build the lambda.
+    ExprResult Call = TokenInjectionHandler->ParseAsExpression(
+        R"c++(
+        [&] consteval {
+           __PTRDIFF_TYPE__ __result = 0;
+           __expansion_stmt_declare_begin_end __begin __end
+           for (; __begin != __end; ++__begin) ++__result;
+           return __result;
+        }()
+      )c++",
+        {
+            {"__expansion_stmt_declare_begin_end", DeclareBeginEnd},
+        },
+        Loc);
+    if (Call.isInvalid() || Call.get()->isTypeDependent())
+      return std::nullopt;
+
     Expr::EvalResult ER;
     SmallVector<PartialDiagnosticAt, 4> Notes;
     ER.Diag = &Notes;
@@ -641,7 +690,6 @@ Sema::ComputeExpansionSize(CXXExpansionStmtPattern *Expansion) {
     // via the built-in '++' on a ptrdiff_t.
     assert(ER.Val.getInt().isNonNegative());
     return ER.Val.getInt().getZExtValue();
-#endif
   }
 
   assert(Expansion->isDestructuring());